bin/requirements.txt
setuptools
rpi_lgpio
paho-mqtt